requirements.txt
git+https://github.com/django-nonrel/django@nonrel-1.5
django-cors-headers==0.13
django-mongodb-engine==0.5.2
djangorestframework==2.4.3
djangotoolbox==1.6.2
facebook-sdk==0.4.0
mongoadmin==0.2
mongodbforms==0.3
mongoengine==0.8.7
pymongo==2.7.2
wsgiref==0.1.2